+++ sqlite3/test/hook/description_test.dart
@@ -57,13 +57,48 @@ void main() {
         return build(args, (input, outputs) async {
           final config = SqliteBinary.forBuild(input) as CompileSqlite;
 
-          expect(config.sourceFile, p.join(d.sandbox, 'native', 'sqlite3.c'));
+          expect(config.sourceFiles, [
+            p.join(d.sandbox, 'native', 'sqlite3.c'),
+          ]);
           expect(config.additionalIncludes, [
             p.join(d.sandbox, 'native', 'include'),
             '/absolute/include',
           ]);
           expect(config.additionalLibraryDirectories, [
             p.join(d.sandbox, 'native', 'lib'),
+          ]);
+        });
+      },
+      check: (_, _) {},
+      extensions: [
+        CodeAssetExtension(
+          targetArchitecture: Architecture.arm64,
+          targetOS: OS.macOS,
+          linkModePreference: LinkModePreference.dynamic,
+          macOS: MacOSCodeConfig(targetVersion: 13),
+        ),
+      ],
+    );
+  });
+
+  test('multiple source files', () async {
+    await testBuildHook(
+      userDefines: PackageUserDefines(
+        workspacePubspec: PackageUserDefinesSource(
+          defines: {
+            'source': 'source',
+            'path': ['native/sqlite3.c', 'native/extra.c'],
+          },
+          basePath: Uri.file(p.join(d.sandbox, 'pubspec.yaml')),
+        ),
+      ),
+      mainMethod: (args) {
+        return build(args, (input, outputs) async {
+          final config = SqliteBinary.forBuild(input) as CompileSqlite;
+
+          expect(config.sourceFiles, [
+            p.join(d.sandbox, 'native', 'sqlite3.c'),
+            p.join(d.sandbox, 'native', 'extra.c'),
           ]);
         });
       },